Add internal protobufs
Relates to #3292

Adds a mechanism for building internal-only protobufs. This is intended
for situations where we use a protobuf internally in Minder, but not as
part of our API contract. No docs will be generated for these protobufs.

As part of this work, add a new protobuf for encrypted data, this
relates to the project in the specified ticket.