Decouple crypto engine from encryption algorithm (#3293)
Fixes #3292

Previously the crypto engine was coupled to the AES-CFB algorithm. This
PR introduces a new EncryptionAlgorithm interface which represents
different cryptographic algorithms. The Engine is now instantiated with
an instance of this algorithm.

Future PRs will make it possible to support more than one algorithm, and
allow the selection of available algorithms to be driven through
configuration.