Initial KeyStore implementation
Relates to #3304

Implement an interface which hides the specifics of loading an
encryption key. This allows us to use a secret store or a key store
directly in future.

Right now, the implementation uses a single key loaded from the config.
In my next PR, I will generalize it to support multiple keys. This will
require a new config structure.