Add provider configuration to the session store (#3348)
* Extend the GetAuthorizationURLRequest protobuf with a config

When GetAuthorizationURLRequest is called, it also creates the session
state for the provider about to be created. Let's extend the message to
include the config.

* Add providerClass to GetAuthorizationURLRequest

At the moment the session state treats the name and the class the same,
which works because we only have github. Let's pass the name and the
state separately.

For backwards compatibility, we still use the provider name as the class
unless set explicitly.

* Pass the config from GetAuthorizationURL to a newly added database column in the session_store table

Store the configuration passed through the GetAuthorizationURL to the
database so it can be retrieved when actually creating the provider.

* Bump database migration version

* Fix tests

* Fix lint

* Bump the migrations again

* Regenerate sqlc after new make bootstrap