Implement AES-256-GCM encryption
Fixes #3317

The GCM mode is recommended by OWASP and other authorities for secret
encryption. As per @jhrozek's suggestion, use the cryptopasta implementation of
AES-256-GCM (see link/copyright in code).

This PR does not actually make use of the new algorithm yet, this will
be done in a future PR.