Remove salt field from EncryptedData
Relates to #3317

Previously EncryptedData had a dedicated salt field. As part of
investigating AES-256-GCM, I have decided to follow the Go crypto
library's convention of prepending the nonce/salt to the ciphertext. As
a result, the salt field is no longer needed.

Since we are still using the hardcoded salt at this time, this will not
impact any data in production.

Also tweaked some unit tests to use the new encrypted data structure
instead of relying on the old DB columns which will be deleted in a
future PR.