Provider classes self-register webhooks
This change makes provider classes self-register webhooks. This means
that a provider class manager now has to implement a registration method
as well as an HTTP handler for the webhook of the given provider.

general webhooks have the following path prefix: `/api/v1/webhook/`.
And, with this change, we now formalize the `/api/v1/webhook/github`
which we had already been using for webhooks for the github provider.

There is also a method of registering "other" handlers, which pertains
to paths that are not the main webhook, but require handling anyway.
GitHub registers two: `ghapp` and `gh-marketplace`. These have the
`/api/v1` prefix, which works with the same paths we already had.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>