Switch to AES-256-GCM encryption for secrets (#3356)
Use AES-256-GCM encryption by default

Fixes #3317

Use GCM encryption as the default encryption algorithm for Minder. At this
point, we now have a hard requirement to use the new EncryptedData structure in
the database - the old columns (for the access token, and for the redirect URL)
do not track the algorithm used. As part of the migration away from the old
columns, I have made the following changes in this PR:

1. Make the old columns nullable
2. Stop writing to the old columns
3. Change all unit tests which relied on the old columns