Add the protobuf struct for generic provider options
In order to actually set the auto-registration of entities, we need to
be able to configure those options.

This patch adds two structures: AutoRegistration which contains a list
of entities to auto-register and ProviderConfig which contains the
AutoRegistration.

The idea behind this is that any provider would be able to use the
ProviderConfig attributes regardless of the provider type. So as an
example, both would be valid:

```
{ "auto_registration": { "enabled": ["repository"] }, "github": { "endpoint": "https://api.github.com" }}
```
and:
```
{ "auto_registration": { "enabled": ["repository"] }, "github-app": { "endpoint": "https://api.github.com" }}
```

and all would require just embedding the ProviderConfig structure into
the provider-specific parser.

Fixes: #3266