Define rule instance table for Minder
Rule instances are currently stored as serialized protobuf structures in
the database, grouped by entity type. A separate join table exists to
track the relationship between a profile-entity and the rule types which
are used. This arrangement leads to some rather awkward code and
database queries making the profile-related code in Minder more complex
than it should be.

This PR adds in a definition of a dedicated rule instance table.
Subsequent PRs will migrate data over to it and then change the codebase
to query this table instead of the current tables.