Remove algorithm from crypto config
Relates to #3315

Previously, it was possible to set a default and fallback algorithm.
This seemed like a good idea at the time but I came to realize the
following:

1) Unlike the key, the algorithm is not an external piece of data - it
   corresponds to something implemented in the codebase. We cannot
   introduce new algorithms without changing the code.
2) After pairing on a key rotation with Ozz, I realized that the
   relationship between the key ID and the algorithm in the config is
   not really intuitive, and is likely to cause confusion which could
   result in us breaking our ability to decrypt secrets.

This PR removes the algorithm from the config and defines the default vs
fallback in the code. As part of this change, the new GCM algorithm
becomes the default for everyone, and this required changing some tests
to use real encryption keys (the old CFB algorithm would accept
arbitrary strings as keys, the GCM cipher expects a base64-encoded
32-byte key).