Treat the provider config as an override, keep the defaults in the code
Since storing the whole config in the database is problematic for a
number of reasons, like hard patching or worse ability to change
defaults, let's change the way we treat the provider configuration in
the database. Instead of storing the full config, we treat the provider
config as an override. To that end, we store a structure with defaults
in the code and make all the fields in the protobuf messages optional
which causes the resulting Go structures to have pointers to fields
instead of direct attributes.

We store the config override verbatim, after just having unmarshalled
and marshalled back to get rid of any extra keys. Then we merge with the
default config on retrieving the provider configuration from the
database.