Recover from panics from the control plane (#3519)
This adds middleware for the GRPC and HTTP handlers so we would only
log panics instead of minder crashing and having to restart. This
ensures that we don't stop serving requests even if something has gone
wrong in our code. We need to start monitoring for panics in the logs
then.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>
Co-authored-by: Don Browne <dmjb@users.noreply.github.com>