Dual write rule instances to new and old tables (#3486)
Relates to #3485

Any time a profile is created or updated, write the rule instances to
both the new and old rule instance tables. Mark the entries in the old
tables as migrated so that we skip over them when we run the migration
process.

Also made def/params field nullable.