Delete profiles before deleting project.
This avoids an issue which Ozz ran into where the `ON CASCADE DELETE`
relationships would attempt to delete rows from rule_type before
deleting from rule_instances. This would fail because there is a foreign
key relation from rule_instances to rule_type which prevents deletion of
in-use rule types.

The old entity_profile_rules table had a foreign key reference to
rule_type marked as `ON DELETE CASCADE`, and relied on logic in the code
to prevent the deletion of rule instances which are in use.