Define evaluation history schemas (#3565)
Fixes #3555

Implements the new database schema for storing evaluation history. Compared with the existing tables for profile/rule status, this will track every single evaluation, and not just the most recent state. This will be used to build an API for querying the history of evaluations for a given rule/entity pair.

Queries and logic to fill the tables will be implemented in the next PR.