Handle the old key when writing configuration, but write the new one (#3575)
Our current configuration key for the GH app provider is "github-app" which
is problematic to use as a key in protobuf. Let's rename the key to
"github_app", but do it in steps - first, this patch ensures that any
new providers are written with underscore, but the old providers
continue to work. As the next step, we're going to migrate the database
records in order to not risk a race where we migrate the database but
the code would still be creating providers with the old key.

Related: #3569

Co-authored-by: Don Browne <dmjb@users.noreply.github.com>