Recover from panics in policy engine
Currently, if there's a panic caused by the evaluation of a policy or
execution of an action, Minder panics and restarts. This is not ideal.
We instead explicitly recover from panics and surface the error up to
the user.

A user will only see that there was an internal error in Minder, but a
relevant log entry will be written.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>