Take into account hierarchy when dealing with rule types
Rule types are usable by projects down the hierarchy, this allows
someone in a child project to create a profile that uses rule types
approved in a parent project. This is very handy since folks with
permission to write profiles don't need to have permissions to write
rule types, they can simply use some already approved and created.

This also adds the constraint that rule type names are unique down the
hierarchy. That is, a user cannot create a rule type name that overrides
a rule type that exists in a parent project.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>