Updating provider config (#3468)
* Protobuf changes for providers update

This merely adds the PatchProvider rpc changes to keep the other patches
in the same PR minimal and don't mix the RPC changes with the logic
changes.

* Extend the ProviderStore with an Update method, so far only allows updating config

The providerStore.Update method only exposes the new config as a
parameter to change. Others might be added as we have a use-case for
them - this is not public API.

* Add a PatchProvider server handler, currently just updating the config

Building atop the providerStore.Update method, we add a PatchProvider
handler that so far only allows updating the config.

* Treat the provider config as an override, keep the defaults in the code

Since storing the whole config in the database is problematic for a
number of reasons, like hard patching or worse ability to change
defaults, let's change the way we treat the provider configuration in
the database. Instead of storing the full config, we treat the provider
config as an override. To that end, we store a structure with defaults
in the code and make all the fields in the protobuf messages optional
which causes the resulting Go structures to have pointers to fields
instead of direct attributes.

We store the config override verbatim, after just having unmarshalled
and marshalled back to get rid of any extra keys. Then we merge with the
default config on retrieving the provider configuration from the
database.