Fall back to `$XDG_CONFIG_PATH/minder/config.yaml` when reading configuration (#3645)
This allows us to have a predictable configuration file when reading the
client minder configuration. The idea is that we could set up
`$XDG_CONFIG_PATH/minder/config.yaml` and that would serve as an entrypoint.

In the future, we could modify this if we want to dynamically change
the current working project.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>