Define evaluation metrics
Add in some basic metrics to track rule/entity evaluations to the
engine. Note that the metrics for alerts and remediations are not wired
in yet since they depend on some changes which will be added in the next
PR.

I have taken a different approach to handling noop metrics than is used
in the codebase so far: instead of defining an interface and separate
real vs noop implementations, I have decided to make use of the noop
metrics handling built into otel: there is now a metrics meter factory
with alternative implementation for noop vs "real" metrics. Any part of
the code which needs to create metrics can use the factory without
caring whether the metrics are actually exported or not.

Relates to: #3556