Added ListEvaluationHistory RPC implementation.
ListEvaluationHistory RPC returns a paginated list of evaluation
events based on the given cursor and filter.

Routines for managing a filter are implemented as well, in a way that
should be reusable within other endpoints supporting filters. This
implementation only allows and-joined predicates. Such predicates
allow only or-joined equality/inequality checks. Simply put, if a
filter entry starts with the exclamation mark, it is added to the
inequality check, while it is added to the equality check
otherwise. Finally, timerange based filtering is supported.

Routines for managing a cursor are implemented as well, but are not
intended to be generic. Cursors are tightly coupled with the
underlying extraction logic and are harder to refactor, and the
additional effort was not considered valuable at this time.

Fixes #3746