Split message handling logic out of Executor (#3730)
Prior to this change, the interface of the Executor directly consumes
messages from a message queue, and then publishes its result to another
queue. This leads to some issues with testing:

1) The tests must be asynchronous in order to simulate event
   publishing. If the tests fail while running `make test`, the test
   hangs until the timeout happens.
2) Errors which take place inside the Executor get swallowed into the
   message and telemetry code, meaning that it's hard to figure out why
   exactly a test has failed.

This PR splits the Executor into two: the ExecutorEventHandler which
handles all interactions with message queues, and the Executor which
contains the logic for entity evaluation minus event handling.

The intention is that the tests for ExecutorEventHandler can use a mock
of the Executor. Since the message handling logic is unlikely to change
often, this should reduce the situations where the async tests fail and
we end up with long hangs in the test. This test cleanup work will be
implemented in a separate PR.