Extend the Profile List database calls to include selectors
Since the profile selectors need to be included when listing profiles,
we need to extend the two profile list calls to list selectors as well.

Because the selectors are stored in a different table, but we still need to
retrieve them in the same Go struct defined in models, we define a new SQL
type in a migration that represents the selectors and tell sqlc to use it in
sqlc.yaml.

This new Go struct needs a Scan method as well to convert raw SQL rows
into structs as well.

Fixes: #3721

Co-authored-by: Michelangelo Mori <mmori@stacklok.com>