ListEvaluationHistory query is generated dynamically.
This is an alternative implementation of the ListEvaluationHistory
endpoint using goqu sql builder library. Generated queries are
equivalent to the one implemented using sqlc.

Some of the code from internal/db/eval_history.go is taken from
sqlc-generated internal/db/eval_history.sql.go in order to minimize
code changes for this POC. Additionally, the generated query can be
further optimized with this approach using data from filters, e.g. by
reducing the number of joins when filtering by entity type.

The big downside is that the amount of glue code to write roughly
doubles because of the need for custom structs and db access layer.