Implement `BoolFromProject` in feature flag code
The `Bool` function for feature flag checks assumes that the context
argument contains all the information needed to do a feature flag
lookup. This is not the case when checking feature flags outside of an
API request context (e.g. in the engine code). The code also uses a User
ID as the targeting key, and it is not clear what the user ID could be
when running asynchronous code.

This adds a second function which takes an explicit project ID argument,
and uses it as a targeting key.