Extend the db-to-pb profile code to include selectors (#3854)
* Extend the ProfileRow interface for selectors

The ProfileRow interface is used to abstract DB rows list with or
without labels that are subsequently passed to
`MergeDatabaseListIntoProfiles`. Let's extend the interface with a
method to get selectors so we can use them in
`MergeDatabaseListIntoProfiles`.

Fixes: #3723

* Add selectors to profile in `MergeDatabaseListIntoProfiles`

`MergeDatabaseListIntoProfiles` creates a protobuf representation of a
profile from a database list. Extend the function with the profile
selectors from the database list.

Related: #3723

* `MergeDatabaseGetIntoProfiles` includes selectors

This will be used in selector handlers.

Related: #3723