CheckHealth RPC now logs error when failing. (#3860)
This change aims to add a little more introspection on why health
check might fail. It's meant to help in scenarios like e.g. when
Minder goes CLBO because of liveness check failures. Currently, it is
hard to tell whether this is because of lack of resources, most likely
CPU, or because it fails to connect to the database.