Print provider config in provider get, if any
We used to only print provider config if there was any on `provider
get`. That's not very user-friendly in case the user enables
auto-registration and then runs provider get to see what had changed.

In the usual case where there's no config, we don't print anything. If
there is configuration, we print key-value pairs for each config option.

The key-value pairs work fine for now since we have a low number of
config options. In the future, once our config gets more complicated, we
should just switch to printing the config as YAML, but for now this
looks awkward and the table gets too vertically long for just one
option.

Fixes: #3816