Fix potential race condition in rule type engine cache (#3918)
This addresses a potential race condition I discovered while working on
another task. If a rule instance with a new rule type is added to a
profile after the rule type cache is populated, but before the profiles
and rule instances are queried, the evaluation will fail because the new
rule type will be missing from the cache.

This changes the logic to query for rule types which are not found in
the cache, and create a new rule type engine instance and cache it if we
get a result.

*Alternatives Considered*

I could simplify this to remove the pre-population and query the
database for each rule type when we encounter it first. This would
simplify the logic, but at the expense of throwing away the performance
gains from loading most of the rules we need in one query.