Added support for cursors to `minder history list`.
This change addes two flags to `minder history list` command, namely
`--cursor`, which lets the user specify the cursor to a specific page
of the list, and `--size`, which lets the user specify the number of
items to retrieve. The two flags can be used independently from one
another.

For obvious reasons, code was also added to print the cursors. I
picked an arbitrary position, but reviewed the style with James to
ensure some degree of uniformity with the rest of the product.

Fixes #3913