Sizes of cursors and pages are integers in REST.
This change addresses the problem of uint64 not being representable in
javascript, leading to a swagger definition of two numbers related to
cursors, namely `size` and `total_records`, having string type rather
than integer. Using a uint32 generates the desired type and is
sufficient for paging purposes.

This is a breaking change, which is why this commit contains changes
to CLI as well. For the record, the feature is currently under feature
flag.

Fixes #3934