Use rule instance table in executor (#3899)
This PR changes the executor to load rule instances from the
`rule_instances` table instead of using the serialized protobufs in
`entity_profiles`.

Using the `rule_instances` table, a new querying method is added to load
all profiles relevant to evaluation, along with the rule instances for
the entity type being evaluated. Two domain-level types -
`ProfileAggregate` and `RuleInstance` are defined to represent the
result of these queries. I have removed a lot of the protobuf usage from
the evaluation code since converting to and from protobufs just adds
complexity, and the new types line up with the values and types expected
by the evaluation code.