Respect the action overrides
PR #3899 accidentally removed a call to GetOnOffState, which allows
individual actions to override the remediation and alerting status in
the profile. This leads to some unusual behaviour, such as attempting to
run the noop alerts and remediations instead of skipping them. This PR
reinstates the calls to GetOnOffState, but changes the function to
accept the ActionOpt status for that action type instead of taking a
pointer to the entire profile.