Add history purge command to minder server cli.
This command is used to manage the life cycle of the history log. The
business requirement is to delete all records older than 30 days
maintaining the most recent one for each entity/rule pair even if
older than 30 days.

Implementing this requirement mandates the processing of the whole set
of records older than 30 days, which cannot be processed in chunks
without creating arbitrary holes in the history. As a first
approximation, the proposed implementation loads all records in RAM,
filters out the ones to keep, and issues a series of deletions of up
to 1000 records each. A test was added to keep track of the record
size. A future improvement would be to spill records to secondary
storage, where we would perform sorting and filtering, but it was
overly complex and unjustified at this point in time.

Fixes #3636