Add `rule_entity_id` to `rule_evaluations`
In order to replace the old evaluation status tables, we need to migrate
any evaluation statuses to the new tables which predate the introduction
of the new evaluation history tables. In order to simplify this
migration, create a foreign key link from `rule_evaluations` to
`evaluation_rule_entities` (the latter is essentially the new equivalent
of the former).

This PR also fixes some issues with the migration introduced in this
previous PR.