Add testcontainers basic tooling.
This change aims to add basic tooling to run custom containers at test
time.

In the interest of time, only support for postgres was added. Routine
`NewPostgresContainer` configures a postgres container to accept
connections to database `minder` with some arbitrarily defined
username and password.

Returned values are a connection string to use for configuring the
storage layer, and a container object that must be terminated by the
caller.