Surface selector check errors to handlers (#4021)
* Distinguish between not implemented and invalid entity for selectors

We had a bug in the CheckSelector method where if we passed in an entity
string representation that doesn't correspond to an entity (aka
garbage), we would convert that to UNSPECIFIED as if the user had
selected 'entity' in the selector.

* Surface selector check errors to handlers

* Fix selector syntax errors in handler tests

Turns out our handler tests used malformed selectors. We didn't know
because we didn't check the syntax previously.