Populate evaluation history tables with old evaluations
In order to replace the old rule status tables, we need to populate the
evaluation history tables with the statuses of rule evaluations which
happened prior to the introduction of the new tables. Once this
migration is complete, it will be possible to change the existing
profile and rule queries to use the evaluation history tables instead of
the old rule evaluation/details tables.