Change profile validation to enforce case insensitive unique rule names
Previously, minder required rule names to be unique within an entity
type, but did so in a case-sensitive manner. Minder's API expects rule
names to be unique in a case-insensitive manner. This changes the
validation logic to prevent new rule instances from being created if
they are not unique in a case insensitive manner.