Drop `rule_entity_id` column from `rule_evaluations`
This was added in a previous PR to simplify a PR. I have since decided
on a simpler approach which uses a boolean `migrated` column instead of
a FK reference. Using an FK reference ended up requiring some disruptive
changes to some of our tests, so I decided to take the simpler approach
instead.