Fix migrated column on rule evaluations
After testing in staging, I found that the migration to set the
`migrated` column did not work as expected. The intention was that any rows in
`rule_evaluations` which have the same combination of rule instance ids
and entity ids should be marked as migrated, as they refer to the same
evaluations. This migration addresses this problem.

Tested against staging (with a rollback)