Fix rule_instance_id column on rule_evaluations
When this column was backfilled originally, the update statement lacked
some of the fields needed to make sure that the join from
rule_evaluations to rule_instances was a 1-1 mapping. This led to the
wrong rule_instance_id being attached to the rule_evaluations table.
Note that this value is not currently used for anything, so the impact
was negligible.

This re-runs the backfill with all fields matching to ensure the correct
ID is used.