Change rule_evaluations upsert to set `migrated` to true
This query is always called after an upsert into the evaluation history
table, so migrated should always be set to true. Due to the lack of the
update in the `ON CONFLICT DO UPDATE` clause, we ran into a situation
where unmigrated rows in rule_evaluations were being implicitly migrated
by the policy migration (i.e. the code was creating a corresponding row
in evaluation_rule_entities) but the migrated flag was still set to
false. The migration script would then attempt to copy the data to the
evaluation_rule_entities table, which would error out because the data
for that rule/entity pair already existed.