Redo migration from PR #4065 (#4129)
Refer to PR #4065 for a description of the change. This was merged in
but had to be rolled back due to a bug. That bug was fixed in PR #4124.
As part of this roll out. Part of migration script 88 needs to be
repeated to fix any rows affected by that bug before applying it.