Handle github repository transfer events. (#4130)
Github webhook events having type `repository` and action
`transferred` are handled by producing an internal event that triggers
an evaluation. Despite being safe, it cannot lead to any useful action
since the github app installation lost permissions to operate on the
repository.

This change makes it so that `repository.transferred` events are
handled as deletions. In order for such deletion to be successful, 403
Forbidden errors on webhook deletions must be handled gracefully, so
that repository deletion can terminate without errors.

This is safe since (b) there's nothing Minder can do about this,
having lost the necessary permissions, and (b) repository registration
removes stale webhooks. The latter point is relevant to fix repository
transfers when auto registration is enabled: in these cases,
repository ownership is migrated before the event is delivered to
Minder, making it impossible to cleanup webhooks, which leads to an
SQL deletion failure, which eventually makes it impossibile to
re-register the repository under the destination project.

Fixes #3274

Co-authored-by: Don Browne <dmjb@users.noreply.github.com>