Check for non-empty length of previous metadata (#4182)
If there is no prior alert state for a rule/entity pair, an empty
db result struct is passed around in the evaluation context. Due to a
change from an inner join to a left join, the field's type changed from
a pointer to a byte slice to just a byte slice. In the empty case, this
means that the alert metadata is represented by an empty byte slice
instead of a nil. Change the code to check for empty length in addition
to nil-ness.