Fix issue with Alert URLs for pull requests and artifacts
The `GetProfileStatusByName` URL returns a bunch of information about
profiles, including the URL of GHSA alerts if one was created. Minder
does not store this URL in the DB, and instead creates it on the fly in
the handler code. This assumed that the repo name and owner was returned
from the `ListRuleEvaluationsByProfileId` SQL query for all entities,
not just repositories.

Since my recent changes removed the link between evaluation history and
repositories for artifacts and pull requests, this logic stopped
working. This works around the problem by querying the DB explicitly.

Ideally, the URL would be stored in the database as part of the alert
metadata. This handles, for example, self-hosted Github, or indeed,
alerts other than GHSAs. I consider that a longer term change which
requires additional changes. This fixes the immediate problem.