Use nil evaluation state if there is no previous evaluation
The rule engine has a copy of the previous evaluation state as part of
its context. Prior to this PR, if there is no previous state for a
rule/entity pair, an empty struct would be passed around. This requires
each part of the code which uses the previous state (particularly
remediations and alerts) to test if the fields in the state struct are
the default value. This problem was highlighted in PR #4089 when the
types of some the fields changed, requiring the code to be changed to
have different default checks.

This PR changes the logic so that a nil pointer is returned if there is
no existing state. This makes the check for a lack of previous state to
be consistent irrespective of the fields needed or their type.

Validated with unit tests locally.