Fix issue with missing evaluation results (#4226)
Previously, an if statement contained multiple clauses, one of which was
missing a `!`. This led to errors happening when they should not have
been triggered. The problem which the error check was detecting was
largely cosmetic, so I have simply removed the if statement. I also
changed the original error handling logic slightly - the code previously
would insert empty evaluation responses into the response if an error
was encountered. Change the logic to avoid creating these empty
statuses.