Decouple minder core from entities.
Webhook handlers should pack all the necessary information to allow
Minder Core to process creations, deletions, and reconciliations of
entities, regardless of their kind. To achieve extensibility, (a)
entities cannot be part of the message topic used for communication,
and (b) messagess cannot be tied to specific entities.

To achieve that, reconciler handlers must dispatch an event to
provider-specific code, shipping the minimal mandatory information
explicitly (i.e. project id and provider id), and entity specific
information in an opaque, provider-specific way.

This change introduces a bit of boilerplate to decouple message
handling from provider-specific code, but does not move the code from
the reconciler to the provider.

Fixes #4247