Handle int64 and uint64 in properties by wrapping it
We started using structpb.Value to hold the payload of our properties,
but since the Value type is expected to be directly serialized to a JSON
value, all numerical values are treated as float64.

Because we also want to store int64 values, we need to treat them
separately, by storing them as objects instead with the integer value
stringified.

Related: #4171