Change the Provider's FetchProperty/FetchAllProperties interface methods to look up by Properties, not just a name
The initial implementation of the `FetchProperty/FetchAllProperties`
methods was very github specific in the sense that it only allowed
looking up the entity by name. Other providers might not have a name and
the provider is best equipped to know what to look by.

Let's send Properties instead and let the provider pick the right ones
or return an error.

Related: #4171