Validate rule type `guidance` as markdown.
We're currently treating `guidance` as a plain string, storing it
unmodified.

This change introduces some controls ensuring that guidance is valid
markdown before storing it in the database. Since any valid UTF-8
string is valid markdown, the check is a bit redundant at the moment
and ensures that it is parseable and renderable as HTML.

Additionally, we limit its size to 10kB, which is far more than enough
given the current rule types have a `guidance` field shorter than 1kB.

Fixes #4286