Render rule `guidance` as markdown in the CLI.
This change adds code to render the `guidance` field of a rule type
definition as markdown. In case it is not possible to render it
properly, it just renders the string as given.

The library used for rendering, `glamour`, internally checks for HTML
tags and strips them completely from the output, ensuring safety.

Fixes #4285