Validate rule `guidance` strictly.
This change uses `bluemonday` library to perform strict validation of
`guidance` field of rule types. In case sanitized input is different
from the input itself, rule type creation/update is rejected returning
a meaningful error message.

The way input is sanitized is in line with the way it is rendered in
the CLI by the `glamour` library.