Take the property service into use in webhook and repository service (#4299)
* Accept transaction from outside in the property service

If we had called these methods in a transaction, we would have
effectivelly created another parallel transaction which might cause
entities to not be referenced correctly.

* Add FilteredCopy and Merge to properties

These will be useful later.

* Add EntityInstance model

This is just a shortcut to avoid passing the properties and project ID
and entity etc around

* Decouple entities from fetching properties

This makes it easier to handle multiple entities inside the github
provider as all that is needed is to implement the PropertyFetcher
interface. It also makes moves the logic around formatting the name and
what properties are operational into the provider to code that is
per-entity.

* Handle operational attributes

minder saves attributes like hook ID which do not come from the upstream
entity. We need to avoid deleting them on updates.

* Add a method to refresh a repository + lazy migration into the github repository service

Related: #4179

* Switch over webhook handling to use property entities

Changes fetchRepo to refresh the repository being processed and return
an EntityInstance model structure instead of a database repo

Related: #4173

* Propagate 404 from the property wrappers as a special error code