Derive profile name from display name in `ProfileService.CreateProfile`
Currently, clients of the API are required to derive the profile name from the display name using various transformations.
This results in duplicated logic across the ecosystem, increasing complexity and potential inconsistencies.

This commit scopes the logic to derive the profile name from the display name to the ProfileService.CreateProfile endpoint.
By centralizing this logic, we reduce redundancy and simplify client code.
Clients of the API are no longer need to implement logic to deriive the profile name when creating a new profile.

Signed-off-by: Kugamoorthy Gajananan <gajananan@gmail.com>